refactor : 사용하지 않는 파일 삭제, 헤더 교체
