--- src/entities/follow/api/rejectFollowButton.ts
@@ -1,13 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-
-import { api } from '@/shared/api';
-
-export const rejectfollowUser = async (id: number): Promise<void> => {
-  await api.private.friend.rejectFollowing(id);
-};
-
-export const useRejectFollow = () => {
-  return useMutation<void, unknown, number>({
-    mutationFn: rejectfollowUser,
-  });
-};
