--- src/features/quiz/model/use-quiz-store.ts
@@ -1,6 +1,0 @@
-import { useQuizContext } from './use-quiz-context';
-
-export const useQuizStore = () => {
-  const { store } = useQuizContext();
-  return store();
-};
