+++ src/pages/_authenticated/_fixed-header-layout/route.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, Outlet } from '@tanstack/react-router';
 
-import Header from '@/widgets/header/ui/Header';
+import Header from '@/widgets/header/ui/header';
 
 export const Route = createFileRoute('/_authenticated/_fixed-header-layout')({
   component: FixedHeaderLayoutComponent,
@@ -9,7 +9,7 @@ export const Route = createFileRoute('/_authenticated/_fixed-header-layout')({
 function FixedHeaderLayoutComponent() {
   return (
     <div className="flex flex-col min-h-screen">
-      <Header className="fixed" />
+      <Header />
       <Outlet />
     </div>
   );
