+++ src/pages/_authenticated/main/_components/main-desktop.tsx
@@ -1,7 +1,7 @@
 import LearningStreakCard from '@/entities/learning/ui/learning-streak-card';
 import MissionCard from '@/entities/mission/mission-card';
 import HeroSection from '@/shared/ui/hero/hero-section';
-import Header from '@/widgets/header/ui/_header';
+import Header from '@/widgets/header/ui/header';
 import RecommendedUnitsSection from '@/widgets/main-page/recommended-units-section';
 import UnitListCard from '@/widgets/main-page/unit-list-card';
 import UserProgressBar from '@/widgets/main-page/user-progress-bar';
