+++ src/pages/_authenticated/my/_profile-layout/route.tsx
@@ -4,7 +4,7 @@ import { useGetMyPageBanner } from '@/shared/api/@generated/mypage-api/mypage-ap
 import { HEADER_HEIGHT } from '@/shared/config/constants';
 import useResponsive from '@/shared/model/use-responsive';
 import PageLayout from '@/shared/ui/layout/page-layout';
-import Header from '@/widgets/header/ui/_header';
+import Header from '@/widgets/header/ui/header';
 import UserProfileCard from '@/widgets/user/ui/user-profile-card';
 import UserTabs from '@/widgets/user/ui/user-tabs';
