+++ src/widgets/header/ui/Header.tsx
@@ -1,86 +1,31 @@
-import { Link } from '@tanstack/react-router';
+import { useLocation } from '@tanstack/react-router';
 
 import { useUserInfo } from '@/entities/sidebar/api/useUserInfo';
-import { tokenManager } from '@/shared/api/config';
-import Logo from '@/shared/assets/icons/logo-gr.svg?react';
-import Profile from '@/shared/assets/icons/profile2.svg?react';
-import { cn } from '@/shared/lib/cn';
-import { PROFILE_COLORS } from '@/shared/lib/ProfileColor';
 
-const HEADER_NAV_LIST: HeaderNavProps[] = [
-  { to: '/main', text: '홈' },
-  { to: '/learning', text: '학습' },
-  { to: '/league', text: '리그' },
-  { to: '/user', text: '사용자' },
-];
+import { DEFAULT_HEADER_NAV_LIST } from '../config/nav';
+import HeaderContent from './header-content';
 
-export default function Header({ className }: { className?: string }) {
-  const { data } = useUserInfo();
+function Header() {
+  const pathname = useLocation({
+    select: (location) => location.pathname,
+  });
 
-  let loginState = <div></div>;
+  const { data, isPending } = useUserInfo();
 
-  if (data) {
-    const profileBgColor = PROFILE_COLORS[data.profileImgNumber as keyof typeof PROFILE_COLORS];
+  if (isPending) return null;
+  if (!data) return null;
 
-    loginState = (
-      <Link to="/user" className="h-full flex gap-4 items-center flex-1 justify-end">
-        <Profile style={{ color: profileBgColor }} className="w-[40px] h-[40px]" />
-        <span className="hidden lg:block text-black font-bold text-xl leading-normal">
-          어서오세요, {data.nickname}님
-        </span>
-        <button
-          type="button"
-          onClick={() => {
-            tokenManager.clearTokens();
-            /** TODO 나중에 수정 */
-            window.location.href = '/';
-          }}
-          className="text-gray-500 text-xl font-bold cursor-pointer"
-        >
-          로그아웃
-        </button>
-      </Link>
-    );
-  }
+  const isMainPage = pathname === '/main';
 
   return (
-    <header
-      className={cn(
-        ' flex items-center px-8 py-4 justify-start h-[var(--header-height)] w-full bg-white z-[110] ',
-
-        className,
-      )}
-    >
-      <Link to="/main" className="mr-3">
-        <Logo className={`h-6`} />
-      </Link>
-      <nav className="flex justify-between gap-8 px-8">
-        {HEADER_NAV_LIST.map((nav) => {
-          return <HeaderNav {...nav} key={nav.text} />;
-        })}
-      </nav>
-      {loginState}
+    <header className="fixed left-0 top-0 z-50 w-full px-15 py-5">
+      <HeaderContent
+        navList={DEFAULT_HEADER_NAV_LIST}
+        profileImageNum={data.profileImgNumber}
+        variant={isMainPage ? 'transparent' : 'solid'}
+      />
     </header>
   );
 }
 
-type HeaderNavProps = {
-  to: string;
-  text: string;
-};
-
-const HeaderNav = ({ to, text }: HeaderNavProps) => {
-  return (
-    <Link
-      to={to}
-      className="text-gray-500 font-bold text-xl leading-6"
-      activeProps={{
-        style: {
-          color: '#222124',
-        },
-      }}
-    >
-      {text}
-    </Link>
-  );
-};
+export default Header;
